+++ application/views/dashboard/settings/admin_login/add.php
@@ -23,6 +23,15 @@
                             </div>
                         </div>
                         <div class="form-group">
+                            <label class="control-label col-md-3 col-sm-3 col-xs-12">User Level</label>
+                            <div class="col-md-9 col-sm-9 col-xs-12">
+                                <select name="level" class="form-control">
+                                    <option value="1">Kasir</option>
+                                    <option value="2">Admin</option>
+                                </select>
+                            </div>
+                        </div>
+                        <div class="form-group">
                             <label class="control-label col-md-3 col-sm-3 col-xs-12">Password</label>
                             <div class="col-md-9 col-sm-9 col-xs-12">
                                 <input type="password" class="form-control" placeholder="password" name="password" required>