+++ application/views/dashboard/settings/admin_login/view.php
@@ -4,7 +4,7 @@
         <div class="col-md-12 col-sm-12 col-xs-12">
             <div class="x_panel">
                 <div class="x_title">
-                    <h2>Admin Login</small></h2>
+                    <h2>User Login</small></h2>
                     <div class="clearfix"></div>
                 </div>
                 <div class="x_content">
@@ -13,6 +13,7 @@
                             <tr>
                                 <th>No</th>
                                 <th>Name</th>
+                                <th>Level</th>
                                 <th>Email</th>
                                 <th>Action</th>
                             </tr>
@@ -26,6 +27,7 @@
                             <tr>
                                 <td><?php echo $i; ?></td>
                                 <td><?php echo $user->name; ?></td>
+                                <td><?php echo user_level($user->level); ?></td>
                                 <td><?php echo $user->email; ?></td>
                                 <td>
                                     <a href="<?php echo base_url('admin_login/details/' . $user->id); ?>"><i class="fa fa-eye"></i></a>