+++ changelog.md
@@ -1,4 +1,5 @@
 ## 1.0.01 - 2018-11-01
 * Initial Commit
 * Add function shop version
-* Add featured filter on laporan page
+* Add featured filter on laporan page.
+* Add featured level on user login.